# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for the API Gateway, BFF (Backend-for-Frontend), and all backend microservices of the ecommerce platform.

    ## Architecture Overview
    The system is composed of the following layers:
    - **API Gateway** (`:8080`) — Single ingress point. Validates JWT access tokens from the `access_token` cookie, enforces role-based access control, and proxies requests to the appropriate downstream service.
    - **BFF Service** (`:8088`) — Aggregates data from multiple services into single, page-optimised responses for the web frontend. All BFF endpoints also flow through the Gateway.
    - **Auth Service** (`:8081`) — Issues and validates JWT access tokens (short-lived) and refresh tokens (long-lived). Both tokens are delivered as HttpOnly cookies.
    - **User Service** (`:8085`) — Manages user profiles and credential changes.
    - **Product Service** (`:8082`) — Product catalogue, category tree, bulk import/export, and presigned S3 upload URLs.
    - **Cart Service** (`:8086`) — Per-user shopping cart and checkout initiation.
    - **Order Service** (`:8083`) — Order lifecycle management.
    - **Payment Service** (`:8087`) — Stripe payment session creation, verification, and webhook ingestion.
    - **Inventory Service** — Stock management, reservation, release, and confirmation as part of the order saga.
...

    ## Authentication
...

    ## Role-Based Access Control
...

...

...

...

...
servers:
...
paths:
...
  /bff/auth/register:
    post:
      tags:
        - BFF
      summary: Register a new user account
      description: |
        Creates a new **user** account (role is always `user` for self-registration).
        On success a verification email is sent to the provided address. The user must
        verify their email before they can log in.

        Returns `409 Conflict` if the email is already registered.

        > **Admin Panel Signup:** To register an admin account directly from the Admin Panel UI, include the `X-Admin-Signup-Secret` header with the secret defined in the backend environment. This bypasses the email verification requirement.

        > **Standard Admin Creation:** Alternatively, use `POST /bff/admin/users` with an existing admin JWT cookie.
      operationId: bffRegister
      security: []
      requestBody:
        $ref: '#/components/requestBodies/RegisterRequest'
      responses:
        '201':
          $ref: '#/components/responses/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/login:
    post:
      tags:
        - BFF
      summary: Login and receive auth cookies
      description: |
        Authenticates a user with email and password. On success, sets four HttpOnly cookies:
        - `__session` — short-lived JWT access token (15 min). **This is the token the API
          Gateway validates on every subsequent request.**
        - `refresh_token` — long-lived token (7 days) used exclusively with `POST /bff/auth/refresh`.
        - `user_id` — the authenticated user's UUID (readable by JS, used by BFF internally).
        - `user_role` — the authenticated user's role (`user` or `admin`).

        Returns `401 Unauthorized` for invalid credentials or unverified email addresses.
      operationId: bffLogin
      security: []
      requestBody:
        $ref: '#/components/requestBodies/LoginRequest'
      responses:
        '200':
          description: Generic success message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Operation completed successfully.

          headers:
            Set-Cookie: &authCookieHeader
              description: HttpOnly auth cookies set on successful authentication.
              schema:
                type: string
              example: __session=eyJ...; HttpOnly; Secure; SameSite=None; Path=/
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/logout:
    post:
      tags:
        - BFF
      summary: Logout and clear auth cookies
      description: |
        Invalidates the current session and clears both `access_token` and `refresh_token`
        cookies on the client. Safe to call even if the user is not currently authenticated.
      operationId: bffLogout
      security: []
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/refresh:
    post:
      tags:
        - BFF
      summary: Refresh the access token using the refresh cookie
      description: |
        Issues a new `access_token` cookie using the `refresh_token` cookie. Call this when an
        API request returns `401 Unauthorized` due to an expired access token. Returns `401` if
        the refresh token is missing, expired, or revoked.
      operationId: bffRefreshToken
      security: []
      responses:
        '200':
          description: Generic success message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Operation completed successfully.

          headers:
            Set-Cookie: *authCookieHeader
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/status:
    get:
      tags:
        - BFF
      summary: Check authentication status
      description: |
        Returns whether the current request is authenticated and, if so, the basic identity of
        the authenticated user (ID, email, role). Useful for initialising client-side auth state
        on page load.
      operationId: bffGetAuthStatus
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current authentication status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/verify-email:
    post:
      tags:
        - BFF
      summary: Verify email address with OTP code
      description: |
        Verifies a user's email address using the one-time code sent during registration.
        The `email` and `code` fields are both required. Returns `404` if no pending
        verification exists for the given email.
      operationId: bffVerifyEmail
      security: []
      requestBody:
        $ref: '#/components/requestBodies/VerifyEmailRequest'
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /bff/admin/users:
    get:
      tags: [Admin]
      summary: List all users (admin)
      operationId: adminListUsers
      description: |
        Returns a paginated list of all registered users (customers and admins).
        Passwords are never included in the response. Results are sorted by
        creation date descending (newest first). Requires admin role.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
        - $ref: '#/components/parameters/AdminPageParam'
        - $ref: '#/components/parameters/AdminPageSizeParam'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            email:
                              type: string
                              format: email
                            name:
                              type: string
                            role:
                              type: string
                              enum: [user, admin]
                            phone_number:
                              type: string
                              nullable: true
                            created_at:
                              type: string
                              format: date-time
                      meta:
                        type: object
                        properties:
                          page:
                            type: integer
                          page_size:
                            type: integer
                          total:
                            type: integer
                          total_pages:
                            type: integer
                  error:
                    type: string
              example:
                success: true
                data:
                  users:
                    - id: e3bc8bbb-aee3-445c-b51a-0b3edee3b613
                      email: admin@example.com
                      name: Admin User
                      role: admin
                      created_at: '2026-03-10T01:46:38Z'
                  meta:
                    page: 1
                    page_size: 20
                    total: 23
                    total_pages: 2
                error: ''
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
    post:
      tags: [Admin]
      summary: Create user (admin)
      operationId: adminCreateUser
      description: |
        Creates a user through the auth-service admin flow.
        Requires admin role and supports creating either `user` or `admin` accounts.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                role:
                  type: string
                  enum: [admin, user]
            example:
              name: Ops Admin
              email: ops-admin@example.com
              password: StrongPass#123
              role: admin
      responses:
        '200': {$ref: '#/components/responses/AdminSuccessResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
...
  /auth/register:
    post:
      tags:
        - Gateway
        - Auth Service
      summary: Register a new user account
      description: |
        Direct Gateway route to the Auth Service registration endpoint. Creates a user account
        and sends a verification email. Equivalent to `POST /bff/auth/register` but without
        BFF aggregation.
      operationId: authRegister
      security: []
      requestBody:
        $ref: '#/components/requestBodies/RegisterRequest'
      responses:
        '201':
          $ref: '#/components/responses/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Gateway
        - Auth Service
      summary: Login and receive auth cookies
      description: |
        Direct Gateway route to the Auth Service login endpoint. Authenticates the user and
        sets `access_token` and `refresh_token` HttpOnly cookies.
      operationId: authLogin
      security: []
      requestBody:
        $ref: '#/components/requestBodies/LoginRequest'
      responses:
        '200':
          description: Generic success message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Operation completed successfully.

          headers:
            Set-Cookie: *authCookieHeader
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/refresh:
    post:
      tags:
        - Gateway
        - Auth Service
      summary: Refresh access token using the refresh cookie
      description: |
        Issues a new `access_token` cookie from a valid `refresh_token` cookie. Call when an
        API request returns `401` due to an expired access token.
      operationId: authRefreshToken
      security: []
      responses:
        '200':
          description: Generic success message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Operation completed successfully.

          headers:
            Set-Cookie: *authCookieHeader
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/status:
    get:
      tags:
        - Gateway
        - Auth Service
      summary: Check authentication status
      description: |
        Returns the authenticated identity (ID, email, role) from the current access token
        cookie. Useful for server-side session hydration.
      operationId: authGetStatus
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Authentication status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/verify-email:
    post:
      tags:
        - Gateway
        - Auth Service
      summary: Verify email address with OTP code
      description: Verifies a user's email using the one-time code delivered during registration.
      operationId: authVerifyEmail
      security: []
      requestBody:
        $ref: '#/components/requestBodies/VerifyEmailRequest'
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── User Service ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        JWT access token stored as an HttpOnly cookie. Issued by the Auth Service on login or
        token refresh. Short-lived (typically 15 minutes). Pass the `admin` scope value in the
        `security` declaration to indicate an endpoint requires the `admin` role.
  parameters:
...
    AdminAuthorizationHeader:
      name: Authorization
      in: header
      required: true
      description: Bearer token header forwarded by gateway.
      schema:
        type: string
    AdminUserIDHeader:
      name: X-User-ID
      in: header
      required: true
      description: User identifier forwarded by API gateway.
      schema:
        type: string
    AdminUserRoleHeader:
      name: X-User-Role
      in: header
      required: true
      description: Role header forwarded by API gateway; must be `admin`.
      schema:
        type: string
        enum: [admin]
    AdminPageParam:
      name: page
      in: query
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    AdminPageSizeParam:
      name: page_size
      in: query
      description: Number of records per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  requestBodies:
    LoginRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginRequest'
    RegisterRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RegisterRequest'
    VerifyEmailRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VerifyEmailRequest'
...
  responses:
    MessageResponse:
      description: Generic success message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Operation completed successfully.
...
    BadRequest:
      description: Bad request — the request body or parameters failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Bad Request
            details: Field 'email' must be a valid email address.

      x-retryable: false
    Unauthorized:
      description: |
        Unauthorized — no valid `access_token` cookie present, or the token has expired.
        Refresh the token via `POST /auth/refresh`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
            details: Access token is missing or has expired.

      x-retryable: false
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not Found
            details: The requested resource could not be found.

      x-retryable: false
    Conflict:
      description: Conflict — the request could not be completed due to a conflict with the current state (e.g. duplicate email on registration).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Conflict
            details: A user with this email address already exists.

      x-retryable: true
    BadGateway:
      description: Bad Gateway — the BFF or Gateway received an invalid response from an upstream microservice. This typically indicates the downstream service is unavailable or returned an unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Bad Gateway
            details: Upstream service is temporarily unavailable.

      x-retryable: true
...
    AdminSuccessResponse:
      description: Standard successful admin response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
...
    AdminBadRequestResponse:
      description: Standard admin bad request response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
    AdminForbiddenResponse:
      description: Standard admin forbidden response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
    AdminNotFoundResponse:
      description: Standard admin not found response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
    AdminInternalErrorResponse:
      description: Standard admin internal error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
  schemas:
...
    MessageResponse:
      type: object
      description: A simple message confirming the outcome of an operation.
      properties:
        message:
          type: string
          example: Operation completed successfully.
    ErrorResponse:
      type: object
      description: Standard error envelope returned for all 4xx and 5xx responses.
      properties:
        error:
          type: string
          description: Short error code or title.
          example: Not Found
        details:
          type: string
          description: Human-readable description of what went wrong.
          example: No product found with the given ID.
...
    AdminMeta:
      type: object
      description: Standard pagination metadata envelope for admin responses.
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
    AdminStandardResponse:
      type: object
      description: Standard BFF admin response shape.
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
          type: object
        error:
          type: string
          example: ''
        meta:
          $ref: '#/components/schemas/AdminMeta'

    # ── Auth ───────────────────────────────────────────────────────────────────
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: jane.doe@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: S3cur3P@ssw0rd
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane.doe@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: S3cur3P@ssw0rd
        role:
          type: string
          enum:
            - user
            - admin
          description: |
            Role assigned to the new account. In production, `admin` accounts should only be
            created by existing administrators or via a seeding script.
          example: user
    VerifyEmailRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
          example: jane.doe@example.com
        code:
          type: string
          description: One-time verification code sent to the user's email.
          example: '847291'
    AuthStatusResponse:
      type: object
      description: Current authentication state for the request.
      properties:
        authenticated:
          type: boolean
          example: true
        user:
          type: object
          description: Present only when `authenticated` is `true`.
          properties:
            id:
              type: string
              format: uuid
              example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
            email:
              type: string
              format: email
              example: jane.doe@example.com
            role:
              type: string
              enum:
                - user
                - admin
              example: user
        timestamp:
          type: string
          format: date-time
          example: '2026-02-25T10:00:00Z'

    # ── User ───────────────────────────────────────────────────────────────────
...